Task done
This task implements the output of information that was received from the API using XMLHttpRequest and Fetch.

When you click on the "Load by JS" button, information is loaded into the left block using the XMLHttpRequest object, and when you click the "Load by Fetch" button, information is loaded using the fetch function.

During the subsequent processing of the right block of information, the user can delete, call the data change form and save the changes. When you click on the "Delete" button, a DELETE request is executed, and when you click on "Save", a PUT request is executed.